modularized modal into a separate AwsModal
