+++ client/components/Sidebar.tsx
@@ -1,35 +1,16 @@
-import React, { Component } from 'react';
-import ReactDOM from 'react-dom';
-import Modal from 'react-modal';
+import React from 'react';
+import AwsModal from './awsModal';
 
 interface SidebarState {
   url: string;
   refresh: any;
   connect: any;
-  openModal: any;
-  closeModal: any;
-  modalIsOpen: boolean;
+  awsOpenModal: any;
+  awsCloseModal: any;
+  awsModalIsOpen: boolean;
   awsInfo: any;
   awsInfoChange: any;
 }
-
-const customStyles = {
-  content: {
-    top: '50%',
-    left: '50%',
-    right: 'auto',
-    bottom: 'auto',
-    marginRight: '-50%',
-    transform: 'translate(-50%, -50%)',
-    backgroundColor: '#2a2b47',
-    border: 'none',
-  },
-  overlay: {
-    backgroundColor: 'rgba(12, 12, 17, 0.75)',
-  },
-};
-
-Modal.setAppElement('#root');
 
 const Sidebar: React.FC<SidebarState> = (props: SidebarState) => {
   return (
@@ -59,62 +40,17 @@ const Sidebar: React.FC<SidebarState> = (props: SidebarState) => {
           width="75"
           height="75"
           viewBox="0 0 24 24"
-          onClick={props.openModal}
+          onClick={props.awsOpenModal}
         >
           <path d="M12 3c-4.006 0-7.267 3.141-7.479 7.092-2.57.463-4.521 2.706-4.521 5.408 0 3.037 2.463 5.5 5.5 5.5h13c3.037 0 5.5-2.463 5.5-5.5 0-2.702-1.951-4.945-4.521-5.408-.212-3.951-3.473-7.092-7.479-7.092z" />
         </svg>
-        <Modal
-          id="aws-modal"
-          isOpen={props.modalIsOpen}
-          onRequestClose={props.closeModal}
-          style={customStyles}
-          contentLabel="Cloud Modal"
-        >
-          <p>CONNECT TO YOUR AWS RDS</p>
-          <form id="aws-modal-form">
-            <input
-              type="user"
-              id="user"
-              placeholder="User"
-              value={props.awsInfo.user}
-              onChange={props.awsInfoChange}
-            />
-            <input
-              type="host"
-              id="host"
-              placeholder="Host"
-              value={props.awsInfo.host}
-              onChange={props.awsInfoChange}
-            />
-            <input
-              type="database"
-              id="database"
-              placeholder="Database"
-              value={props.awsInfo.database}
-              onChange={props.awsInfoChange}
-            />
-            <input
-              type="password"
-              id="password"
-              placeholder="Password"
-              value={props.awsInfo.password}
-              onChange={props.awsInfoChange}
-            />
-            <input
-              type="port"
-              id="port"
-              placeholder="Port"
-              value={props.awsInfo.port}
-              onChange={props.awsInfoChange}
-            />
-            <button className="primary-button" onClick={props.connect}>
-              Connect
-            </button>
-          </form>
-          <button className="secondary-button" onClick={props.closeModal}>
-            Cancel
-          </button>
-        </Modal>
+        <AwsModal
+          modalIsOpen={props.awsModalIsOpen}
+          awsInfo={props.awsInfo}
+          awsInfoChange={props.awsInfoChange}
+          connect={props.connect}
+          closeModal={props.awsCloseModal}
+        />
         <svg
           id="information-button"
           xmlns="http://www.w3.org/2000/svg"
